Make it possible to force the encoding.
If libxml2 cannot determine the encoding of the HTML document, it will
assume a default of iso-8859-1. This can be problematic and we need a
way to force the encoding.
For now, we let libxml2 do its thing but make it possible for the admin
to supply an encoding. The only sane value for now is "__double" which
decodes UTF-8 documents that have been double-encoded (because they were
interpreted as iso-8859-1 first, then encoded into utf-8 when returned).